feat: provide diff-rename by exposing LSP TextEdits conversion functions
